requirements.txt
gradio==4.19.2
groq>=0.1.0
python-dotenv==1.0.1
huggingface-hub>=0.14.1
manim==0.19.0
flask==3.0.2
flask-cors==4.0.0
